+++ tests/logic/schemas/test_validate_schema.py
@@ -386,6 +386,36 @@ def test_validate_text_valid_multiline_type():
     validate_schema(wrap_into_basic_schema(schema))
 
 
+def test_validate_text_with_placeholder():
+    schema = {
+        'title': 'Example',
+        'type': 'text',
+        'placeholder': 'Placeholder'
+    }
+    validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_text_with_invalid_placeholder():
+    schema = {
+        'title': 'Example',
+        'type': 'text',
+        'placeholder': 1
+    }
+    with pytest.raises(ValidationError):
+        validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_text_with_placeholder_and_choices():
+    schema = {
+        'title': 'Example',
+        'type': 'text',
+        'placeholder': 'Placeholder',
+        'choices': ['A', 'B', 'C']
+    }
+    with pytest.raises(ValidationError):
+        validate_schema(wrap_into_basic_schema(schema))
+
+
 def test_validate_datetime_schema():
     schema = {
         'title': 'Example',
@@ -538,6 +568,27 @@ def test_validate_quantity_schema_missing_key():
     schema = {
         'title': 'Example',
         'type': 'quantity'
+    }
+    with pytest.raises(ValidationError):
+        validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_quantity_schema_with_placeholder():
+    schema = {
+        'title': 'Example',
+        'type': 'quantity',
+        'units': 'm',
+        'placeholder': 'Placeholder'
+    }
+    validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_quantity_schema_with_invalid_placeholder():
+    schema = {
+        'title': 'Example',
+        'type': 'quantity',
+        'units': 'm',
+        'placeholder': 1
     }
     with pytest.raises(ValidationError):
         validate_schema(wrap_into_basic_schema(schema))
@@ -1033,6 +1084,29 @@ def test_validate_object_reference_schema():
     validate_schema(wrap_into_basic_schema(schema))
 
 
+def test_validate_object_reference_schema_with_action_id():
+    schema = {
+        'title': 'Example',
+        'type': 'object_reference',
+        'note': 'Example Note',
+        'action_type_id': ActionType.SAMPLE_CREATION,
+        'action_id': 1
+    }
+    validate_schema(wrap_into_basic_schema(schema))
+
+
+def test_validate_object_reference_schema_with_invalid_action_id_type():
+    schema = {
+        'title': 'Example',
+        'type': 'object_reference',
+        'note': 'Example Note',
+        'action_type_id': ActionType.SAMPLE_CREATION,
+        'action_id': '1'
+    }
+    with pytest.raises(ValidationError):
+        validate_schema(wrap_into_basic_schema(schema))
+
+
 def test_validate_object_reference_schema_with_invalid_note():
     schema = {
         'title': 'Example',
@@ -1091,7 +1165,7 @@ def test_validate_object_reference_schema_with_unknown_property():
     schema = {
         'title': 'Example',
         'type': 'object_reference',
-        'action_id': 0
+        'object_id': 0
     }
     with pytest.raises(ValidationError):
         validate_schema(wrap_into_basic_schema(schema))